requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
asgiref==3.11.1
    # via
    #   django
    #   django-browser-reload
certifi==2026.4.22
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via requests
colorama==0.4.6
    # via tqdm
cryptography==48.0.0
    # via google-auth
dj-database-url==3.1.2
    # via smart-care (pyproject.toml)
django==6.0.3
    # via
    #   smart-care (pyproject.toml)
    #   dj-database-url
    #   django-browser-reload
django-browser-reload==1.21.0
    # via smart-care (pyproject.toml)
django-cleanup==9.0.0
    # via smart-care (pyproject.toml)
django-environ==0.13.0
    # via smart-care (pyproject.toml)
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.25.2
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.196.0
    # via google-generativeai
google-auth==2.52.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.4.0
    # via google-api-python-client
google-generativeai==0.8.6
    # via smart-care (pyproject.toml)
googleapis-common-protos==1.75.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.80.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
gunicorn==25.1.0
    # via smart-care (pyproject.toml)
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.13
    # via requests
packaging==26.0
    # via gunicorn
pillow==12.1.1
    # via smart-care (pyproject.toml)
proto-plus==1.28.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.6
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.12
    # via smart-care (pyproject.toml)
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.13.4
    # via google-generativeai
pydantic-core==2.46.4
    # via pydantic
pyparsing==3.3.2
    # via httplib2
requests==2.33.1
    # via
    #   google-api-core
    #   stripe
sqlparse==0.5.5
    # via django
stripe==15.1.0
    # via smart-care (pyproject.toml)
tqdm==4.67.3
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   google-generativeai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   stripe
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via django
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.7.0
    # via requests
whitenoise==6.12.0
    # via smart-care (pyproject.toml)
